requirements.txt
Flask
Celery
Pillow
requests
amqp
